[JE2-141] 1. Added order status in constants.ts 2. Function getOrderStatusColor replaced in orders/utils 3. All functions have been moved out of the order details 4. A ProductsTable component has been created that takes an array of products and displays them in a simple table 5. Refactored StatusOrderCell now receives the color based on the status from the getOrderStatusColor function and the status array from the constants file
