+++ frontend/src/constants.ts
@@ -34,3 +34,5 @@ export const passwordStrengthOptions: Options<string> = [
     minLength: 8,
   },
 ]
+
+export const OrderStatus = ['в сборке', 'в пути', 'доставлен']
